Added simple voicemail detection example, and complex one with multiple pipelines to the starter image
